requirements.txt
pandas
Django==1.11.8
uwsgi
gunicorn
django-filter
djangorestframework>=3.5
PyMySQL
Pillow
redis
bson
pymongo
pysnmp
xlwt
xlrd
python-consul
GRR-M2Crypto
django-cors-headers
dulwich
funky
celery==3.1.18
django-celery
celery-with-redis
kombu
python-jenkins
channels
graphic-verification-code